requirements.txt
pytest~=6.2.5
numpy~=1.21.4
pandas~=1.3.4
dataclasses~=0.8
jinja2~=3.0.3
python-dotenv~=0.19.2
requests~=2.26.0